# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Strata — Unified Asset & Operations Brain API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents
      parameters:
        - name: docType
          in: query
          schema:
            $ref: "#/components/schemas/DocumentType"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/DocumentStatus"
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Upload and ingest a document
      description: >
        Accepts a base64-encoded file payload, runs it through the ingestion
        pipeline (OCR / vision extraction, chunking, entity extraction) and
        returns the created document. Processing happens synchronously.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a document with its chunks and extracted entities
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /compliance/gaps:
    get:
      operationId: listComplianceGaps
      tags: [compliance]
      summary: List detected compliance gaps
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ComplianceGapStatus"
        - name: severity
          in: query
          schema:
            $ref: "#/components/schemas/ComplianceSeverity"
      responses:
        "200":
          description: List of compliance gaps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ComplianceGap"
  /compliance/gaps/{id}:
    patch:
      operationId: updateComplianceGap
      tags: [compliance]
      summary: Update a compliance gap's status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplianceGapUpdate"
      responses:
        "200":
          description: Updated gap
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceGap"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /drawings:
    get:
      operationId: listDrawings
      tags: [drawings]
      summary: List digitized P&ID / engineering drawings
      responses:
        "200":
          description: List of drawings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PidDrawing"
    post:
      operationId: createDrawing
      tags: [drawings]
      summary: Upload a P&ID / engineering drawing and digitize it via computer vision
      description: >
        Accepts a base64-encoded drawing image (or PDF), runs it through a Gemini
        vision digitization pass (OCR tag reading + symbol classification + line
        tracing), and returns the created drawing record. Processing happens
        synchronously.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DrawingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PidDrawing"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /drawings/{id}:
    get:
      operationId: getDrawing
      tags: [drawings]
      summary: Get a digitized drawing with its extracted components and connections
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Drawing detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PidDrawingDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteDrawing
      tags: [drawings]
      summary: Delete a digitized drawing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    DocumentType:
      type: string
      enum:
        - safety_procedure
        - equipment_manual
        - inspection_report
        - regulation
        - maintenance_record
        - other
    DocumentStatus:
      type: string
      enum:
        - processing
        - ready
        - failed
    Document:
      type: object
      required:
        [id, title, docType, status, fileName, mimeType, summary, createdAt, updatedAt]
      properties:
        id: { type: integer }
        title: { type: string }
        docType:
          $ref: "#/components/schemas/DocumentType"
        status:
          $ref: "#/components/schemas/DocumentStatus"
        fileName: { type: string }
        mimeType: { type: string }
        summary: { type: ["string", "null"] }
        processingError: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    DocumentChunk:
      type: object
      required: [id, documentId, chunkIndex, content]
      properties:
        id: { type: integer }
        documentId: { type: integer }
        chunkIndex: { type: integer }
        content: { type: string }
    DocumentDetail:
      allOf:
        - $ref: "#/components/schemas/Document"
        - type: object
          required: [extractedText, chunks, entities]
          properties:
            extractedText: { type: ["string", "null"] }
            chunks:
              type: array
              items:
                $ref: "#/components/schemas/DocumentChunk"
            entities:
              type: array
              items:
                $ref: "#/components/schemas/Entity"
    DocumentInput:
      type: object
      required: [title, docType, fileName, mimeType, fileBase64]
      properties:
        title: { type: string, minLength: 1 }
        docType:
          $ref: "#/components/schemas/DocumentType"
        fileName: { type: string, minLength: 1 }
        mimeType: { type: string, minLength: 1 }
        fileBase64: { type: string, minLength: 1 }
    EntityType:
      type: string
      enum:
        - equipment
        - procedure
        - regulation
        - personnel
        - location
        - hazard
        - parameter
    Entity:
      type: object
      required: [id, entityType, name, description, documentId]
      properties:
        id: { type: integer }
        entityType:
          $ref: "#/components/schemas/EntityType"
        name: { type: string }
        description: { type: ["string", "null"] }
        documentId: { type: ["integer", "null"] }
...
    ComplianceSeverity:
      type: string
      enum: [critical, high, medium, low]
    ComplianceGapStatus:
      type: string
      enum: [open, resolved]
    ComplianceGap:
      type: object
      required:
        [id, ruleId, ruleTitle, severity, status, description, recommendation, documentId, createdAt]
      properties:
        id: { type: integer }
        ruleId: { type: integer }
        ruleTitle: { type: string }
        severity:
          $ref: "#/components/schemas/ComplianceSeverity"
        status:
          $ref: "#/components/schemas/ComplianceGapStatus"
        description: { type: string }
        recommendation: { type: string }
        documentId: { type: ["integer", "null"] }
        createdAt: { type: string, format: date-time }
    ComplianceGapUpdate:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/ComplianceGapStatus"
...
    PidDrawingStatus:
      type: string
      enum: [processing, ready, failed]
    PidComponentType:
      type: string
      enum: [valve, pump, vessel, compressor, instrument, sensor, pipe_segment, other]
    PidDrawing:
      type: object
      required: [id, title, status, fileName, mimeType, summary, createdAt, updatedAt]
      properties:
        id: { type: integer }
        title: { type: string }
        status:
          $ref: "#/components/schemas/PidDrawingStatus"
        fileName: { type: string }
        mimeType: { type: string }
        summary: { type: ["string", "null"] }
        processingError: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    PidComponent:
      type: object
      required: [id, drawingId, tag, componentType, label, description]
      properties:
        id: { type: integer }
        drawingId: { type: integer }
        tag: { type: string }
        componentType:
          $ref: "#/components/schemas/PidComponentType"
        label: { type: string }
        description: { type: ["string", "null"] }
    PidConnection:
      type: object
      required: [id, drawingId, fromComponentId, toComponentId, lineType, description]
      properties:
        id: { type: integer }
        drawingId: { type: integer }
        fromComponentId: { type: integer }
        toComponentId: { type: integer }
        lineType: { type: string }
        description: { type: ["string", "null"] }
    PidDrawingDetail:
      allOf:
        - $ref: "#/components/schemas/PidDrawing"
        - type: object
          required: [components, connections]
          properties:
            components:
              type: array
              items:
                $ref: "#/components/schemas/PidComponent"
            connections:
              type: array
              items:
                $ref: "#/components/schemas/PidConnection"
    DrawingInput:
      type: object
      required: [title, fileName, mimeType, fileBase64]
      properties:
        title: { type: string, minLength: 1 }
        fileName: { type: string, minLength: 1 }
        mimeType: { type: string, minLength: 1 }
        fileBase64: { type: string, minLength: 1 }
...