# lib/api-zod/src/generated/types/documentDetail.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Strata — Unified Asset & Operations Brain API
 * OpenAPI spec version: 0.1.0
 */
import type { Document } from './document';
import type { DocumentChunk } from './documentChunk';
import type { Entity } from './entity';

export type DocumentDetail = Document & ({
  /** @nullable */
  extractedText: string | null;
  chunks: DocumentChunk[];
  entities: Entity[];
});