# lib/api-zod/src/generated/types/drawingInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Strata — Unified Asset & Operations Brain API
 * OpenAPI spec version: 0.1.0
 */

export interface DrawingInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  fileName: string;
  /** @minLength 1 */
  mimeType: string;
  /** @minLength 1 */
  fileBase64: string;
}